Move AllowNonAsciiHeaders to new class to fix build
